Fix deploy script
When I rearranged the docker images, I forgot to update the mount volume
locations. That meant that the images weren't being uploaded to a host
volume, so all the images disappeared when we restarted the docker
container (by deploying, for example).

That means that all the images we uploaded in the last couple of days
are probably permanently missing.

Oops. :sob: